+++ packages/middlewares/src/index.ts
@@ -1,2 +1,4 @@
+export { excludedMiddleware } from "./excluded";
+export { homeMiddleware } from "./home";
 export { linksMiddleware } from "./links";
 export { pipeMiddleware } from "./pipe";
